+++ R/backend_future.R
@@ -106,7 +106,8 @@ ft_launch_worker <- function(target, meta, protect, config) {
       ),
       globals = globals,
       label = target,
-      resources = as.list(spec[[target]]$resources)
+      resources = as.list(spec[[target]]$resources),
+      seed = TRUE
     ),
     target = target
   )