+++ NEWS.md
@@ -5,6 +5,8 @@
 * Fix defunct functions error message when using namespace (#1310, @malcolmbarrett).
 * Preserve names of list elements in `.data` in DSL (#1323, @shirdekel).
 * Use `identical()` to compare file hashes (#1324, @shirdekel).
+* Set `seed = TRUE` in `future::future()`.
+* Manually relay warnings when `parallelism = "clustermq"` and `caching = "worker"` (@richardbayes).
 
 # Version 7.12.5